requirements.txt
google-api-python-client==1.7.9
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
requests~=2.31.0
numpy~=1.26.1
scipy~=1.11.3
Pillow~=10.0.1
google~=3.0.0
lyricsgenius~=3.0.1